+++ src/core/include/edge.h
@@ -51,7 +51,7 @@ public:
     const Node& neighbour() const;
 
     const Attributes* attrs() const;
-    const Value& attr(size_t id) const;
+    const Value& attr(int id) const;
     Value attr(const QString& name, Value defaultValue=Value()) const;
 
     void setAttr(const int id, const Value& value);
